remove okexf and unify okex
