+++ vnpy/trader/gateway/okexGateway/text.py
